+++ src/app/app.routes.ts
@@ -1,3 +1,6 @@
 import { Routes } from '@angular/router';
+import { TodoComponent } from './todo/todo.component';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  {path: '', component: TodoComponent},
+];
