+++ src/app/todo/todo.component.scss
@@ -1,0 +1,70 @@
+.col-md {
+  width: 20%;
+}
+
+textarea {
+  resize: none;
+}
+
+.card-offset {
+  margin-bottom: 32px;
+}
+
+.table {
+  margin-bottom: 0px;
+}
+
+.todo-box {
+  cursor: move;
+  background: white;
+  width: 100%;
+  padding: 10px;
+  overflow: hidden;
+
+  td {
+    align-content: center;
+  }
+}
+
+.wrapper-empty {
+  opacity: 0.5;
+  text-align: center;
+
+  td {
+    height: 200px;
+  }
+
+  mat-icon {
+    align-content: center;
+  }
+
+  .mat-icon.mat-icon-inline {
+    font-size: 96px;
+    line-height: normal;
+  }
+}
+
+
+/* Material angular drag and drop cdk */
+.cdk-drag-placeholder {
+  opacity: 0;
+}
+
+/* Animate an item that has been dropped. */
+.cdk-drag-animating {
+  transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+/* Animate items as they're being sorted. */
+.todo-list.cdk-drop-list-dragging .todo-box:not(.cdk-drag-placeholder) {
+  transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.cdk-drag-preview {
+  box-sizing: border-box;
+  height: 100%;
+  border-radius: 4px;
+  box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
+              0 8px 10px 1px rgba(0, 0, 0, 0.14),
+              0 3px 14px 2px rgba(0, 0, 0, 0.12);
+}
