+++ src/styles.scss
@@ -3,4 +3,13 @@
 /* Importing Bootstrap SCSS file. */
 // @import 'bootstrap/scss/bootstrap';
 
-@import "bootstrap/dist/css/bootstrap.css"
+@import "bootstrap/dist/css/bootstrap.css";
+
+html, body {
+  height: 100%;
+}
+
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
